backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
python-valve
pyyaml
aiofiles
itsdangerous